Split Up Special Case Handlers
To make them easier to find and order, and make the import a little bit
cleaner with an index file.
